SAS-332 Dividido el archivo de manejo de efectos para cada carta en particular
