+++ src/core/game_logic/effects/flamethrower_effect.py
@@ -1,0 +1,36 @@
+"""Flamethrower effect."""
+
+from typing import Optional
+
+from core.game_logic.effects.nothing_effect import nothing_effect
+from core.game_logic.game_action import ActionType
+from core.game_logic.game_action import GameAction
+from models.game import Game
+from pony.orm import db_session
+
+def flamethrower_effect(id_game: int, target: Optional[int]) -> GameAction:
+    """Flamethrower effect."""
+    with db_session:
+        game = Game[id_game]
+
+        if game.current_phase != "Defense":
+            raise ValueError("You can't use this card in this phase.")
+        if target is None:
+            raise ValueError("You must select a target.")
+        if game.players.select(id=target).count() == 0:
+            raise ValueError("Target doesn't exists.")
+        if not game.players.select(id=target).first().alive:
+            raise ValueError("Target is dead.")
+
+        return GameAction(action=ActionType.KILL, target=[target])
+
+
+def no_barbecues_effect(id_game: int) -> GameAction:
+    """No barbecues effect."""
+    with db_session:
+        game = Game[id_game]
+
+        if game.current_phase != "Defense":
+            raise ValueError("You can't use this card in this phase.")
+
+        return nothing_effect(id_game)
